Generation ok, back to generated ~ok
Generation working fine, back to parameters done on penelope, works fine
as long as for f and T but inspectors are underestimated and I.NO/ P.NO
negative influence is overestimated. The new, stronger of the priors on
cofactors and f are not changing anything in term of estimation on the
real data.